+++ CHANGELOG.md
@@ -4,6 +4,7 @@ TimesheetPlus - Changelog
 V1.1
 ----
 * Added to resume: When user entered task-number instead entry ID, the last entry of that task is suggested to be resumed
+* Bugfix: Viewing empty timesheet w/ filter was stuck in loop
 
 
 v1.0